+++ tests/save_testref.py
@@ -1,9 +1,10 @@
 """This script allows to create a reference response set for mijnbibliotheek.
 
-When ran, it create reference files, which can be used in the mijnbibliotheek 
-tests as expected data. When the files do not exist, the idea is that the 
+When ran, it create reference files, which can be used in the mijnbibliotheek
+tests as expected data. When the files do not exist, the idea is that the
 relevant tests will be skipped.
 """
+
 import configparser
 import pickle
 import sys